+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - A `callPackage` function to the C platforms.
+- Support for terraform plugins in terraform components.
 
 ## [5.4.0] - 2026-04-23